Revert "Revert changes in the program crate"
This reverts commit 854b07f13767478a684558c50dd9f577807ab4b6.
