+++ program/Cargo.toml
@@ -13,7 +13,7 @@ rust-version = "1.81.0"                          # solana platform-tools rust ve
 include = ["src/**/*", "README.md"]
 
 [package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu", "wasm32-unknown-unknown"]
+targets = ["x86_64-unknown-linux-gnu"]
 all-features = true
 rustdoc-args = ["--cfg=docsrs"]
