+++ sdk-wasm-js/Cargo.toml
@@ -10,8 +10,16 @@ license = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
+solana-instruction = { workspace = true, features = ["std"] }
 solana-program = { workspace = true }
 solana-sdk = { workspace = true }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+console_error_panic_hook = { workspace = true }
+console_log = { workspace = true }
+wasm-bindgen = { workspace = true }
+log = { workspace = true }
+js-sys = { workspace = true }
 
 [lib]
 crate-type = ["cdylib"]
